requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.0
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.3
six==1.16.0
tzdata==2023.3